refactor: customView 패키지 이동
