+++ android/app/src/main/java/com/created/team201/presentation/studyDetail/custom/MultiPickerChangeListener.kt
@@ -1,4 +1,4 @@
-package com.created.team201.presentation.updateStudy.custom
+package com.created.team201.presentation.studyDetail.custom
 
 interface MultiPickerChangeListener {
     fun onLeftChange(value: Int)
