+++ android/app/src/main/java/com/created/team201/presentation/studyDetail/custom/decorator/TodayDecorator.kt
@@ -1,4 +1,4 @@
-package com.created.team201.presentation.updateStudy.custom.decorator
+package com.created.team201.presentation.studyDetail.custom.decorator
 
 import android.text.style.ForegroundColorSpan
 import com.prolificinteractive.materialcalendarview.CalendarDay
