+++ android/app/src/main/res/layout/dialog_study_informaion.xml
@@ -57,7 +57,7 @@
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/tv_study_information_title" />
 
-            <com.created.team201.presentation.updateStudy.custom.IconTextButton
+            <com.created.team201.presentation.studyDetail.custom.IconTextButton
                 android:id="@+id/itb_study_information_capacity"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -70,7 +70,7 @@
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/view_study_information_header_divider" />
 
-            <com.created.team201.presentation.updateStudy.custom.IconTextButton
+            <com.created.team201.presentation.studyDetail.custom.IconTextButton
                 android:id="@+id/itb_study_information_start_date"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -83,7 +83,7 @@
                 app:layout_constraintStart_toEndOf="@id/itb_study_information_capacity"
                 app:layout_constraintTop_toBottomOf="@id/view_study_information_header_divider" />
 
-            <com.created.team201.presentation.updateStudy.custom.IconTextButton
+            <com.created.team201.presentation.studyDetail.custom.IconTextButton
                 android:id="@+id/itb_study_information_period"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -96,7 +96,7 @@
                 app:layout_constraintStart_toEndOf="@id/itb_study_information_start_date"
                 app:layout_constraintTop_toBottomOf="@id/view_study_information_header_divider" />
 
-            <com.created.team201.presentation.updateStudy.custom.IconTextButton
+            <com.created.team201.presentation.studyDetail.custom.IconTextButton
                 android:id="@+id/itb_study_information_cycle"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
